fixing CORS Filter
