--- belajar-nfc-server/src/main/java/belajar/nfc/interceptor/CorsInterceptor.java
@@ -1,20 +1,0 @@
-package belajar.nfc.interceptor;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
-
-public class CorsInterceptor  extends HandlerInterceptorAdapter {
-
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        String origin = request.getHeader("Origin");
-        response.addHeader("Access-Control-Allow-Origin", origin);
-        if (request.getHeader("Access-Control-Request-Method") != null && "OPTIONS".equals(request.getMethod())) {
-            response.addHeader("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE");
-            response.addHeader("Access-Control-Allow-Headers", "Content-Type");
-            response.addHeader("Access-Control-Max-Age", "1");
-        }
-        return true;
-    }
-}
