+++ belajar-nfc-server/src/main/resources/client.properties
@@ -1,0 +1,1 @@
+allowedHosts=http://localhost:9000,http://127.0.0.1:9000
